Initial Implementation
Implemented the algorithm to generate the initial score matrix and fills it using the Needleman-Wunsch Algorithm. Currently, it has some possible bugs since it always pics the position [0][0] for the first equation, instead of picking the value that is diagnaly upwards to the left. This could be the reason why the final matrix is not the same as the one in the document.